+++ src/main/java/com/learn/spring/boot/inventory/model/Inventory.java
@@ -1,5 +1,7 @@
 package com.learn.spring.boot.inventory.model;
 
+import com.sun.istack.NotNull;
+
 import lombok.Data;
 
 /**
@@ -9,13 +11,13 @@ import lombok.Data;
  */
 @Data
 public class Inventory {
-	
+	@NotNull
 	private int id;
-
+	@NotNull
 	private String itemName;
-	
+	@NotNull
 	private String quantity;
-	
+	@NotNull
 	private String price;
 
 }